bug fix: make range cursors point to a single non-term:
when recursively retrieving translation rules that handle slot cursors,
clean up range cursors that point to a non-terminal+terminal combo
in such a way that it points to a non-terminal only.
happens with cases such as `li = [0] * N`, whereby the range cursor
for `0` inside `[0]` is `(parent_ast, 3, 5)`
where 3 is idx of `0` and 4 is idx of `]`.