incomplete fix to a regression caused by f40e34e:
Does not allow collecting choicable nodes from functions.
In GFG, statements to be validated are placed into a function
f_gold, from which choicable nodes are collected.
Add a flag to control whether to skip function_definitions or not.
Incomplete, as does not fix the original issue f40e34e solved.